Add events.
An Interest model has been created along with basic routes and a 
controller to get all categories, a particular one, or create one.

I'm not sure about the populations on the queries -- they will probably 
change when a view on the front end is designed.

Besides, how to upload and save images must be investigated, as they 
will make the views much nicer!