Allow events edition.
Similarly to how users edition was handled, this gets rid of the `NewEventForm` in favor of a generic `EventForm` which handles both event creation and edition.

A user can only edit events created by themself, and admin users can edit any event.

I'm a bit worried about how will this perform when the interests grow large enough. Maybe a dedicated tab for interests would make more sense? This is what I plan to do with users, so maybe doing it with events too would be a good way to go.