Refactor the `UserForm` and allow users' interests selection.
This extracts all the state information from the `UserForm` to the parent component, the recently introduced `UserTabs`. This way, it can easily be shared and stay in sync with the new `UserInterestsForm`, which finally allows users to select their interests!

Because of all the refactoring, I decided to bring back the `SignUpForm`, separating it from the `UserForm`, as everything looks cleaner this way. Maybe the `LogInForm` could be merged into this one, but I don't think it's worth it.