Add a script with tentative interests and categories.
This includes a script that populates the database with a tentative list of interests and categories. The list has been crafted using https://interestexplorer.io/facebook-interests-list/ as a starting point, and retrieving the icons from https://www.flaticon.com/.

Having and maintaining a complete list of interests and categories, together with their icons, probably deserves a separate repository.